+++ Sample Code/Shared/LYRMockContentStore.h
@@ -18,6 +18,11 @@
 @property (nonatomic) NSString *authenticatedUserID;
 
 /**
+ @abstract Defaults to `YES`. If set to `NO`, the content store will not broadcast mock change notifications.
+ */
+@property (nonatomic) BOOL shouldBroadcastChanges;
+
+/**
  @abstract Singleton Accessor for the mock content store
  */
 + (id)sharedStore;
@@ -65,6 +70,8 @@
 - (void)deleteMessage:(LYRMessageMock *)message;
 
 - (LYRMessageMock *)messageForIdentifier:(NSURL *)identifier;
+
+- (NSOrderedSet *)allMessages;
 
 //-------------------------
 // Querying