+++ Tests/LYRUIMessageCollectionViewCellTest.m
@@ -1,0 +1,31 @@
+//
+//  LYRUIMessageCollectionViewCellTest.m
+//  LayerUIKit
+//
+//  Created by Kevin Coleman on 1/26/15.
+//
+//
+
+#import <UIKit/UIKit.h>
+#import <XCTest/XCTest.h>
+#import "LYRUITestInterface.h"
+
+@interface LYRUIMessageCollectionViewCellTest : XCTestCase
+
+@end
+
+@implementation LYRUIMessageCollectionViewCellTest
+
+- (void)setUp
+{
+    [super setUp];
+}
+
+- (void)tearDown
+{
+    [super tearDown];
+}
+
+
+
+@end